fixed time

+++ src/metatrader_balance/metatrader_balance.rs
@@ -36,6 +36,6 @@ pub struct MetatraderBalanceBodySbModel {
     #[prost(enumeration = "BalanceEventType", tag = "8")]
     pub event_type: i32,
 
-    #[prost(int32, tag = "9")]
-    pub time: i32,
+    #[prost(int64, tag = "9")]
+    pub time: i64,
 }
